Added hourly load profile data
Added end use hourly load data from several sources. Please read the
associated Read Me. The idea is to simulate a few of such loads on an
IEEE reference bus for our project. This data is meant to be a backup
in case the KRC data is not to our liking, for example too choppy etc.